product and reported ads backend

+++ aclassdeal/src/main/java/com/anaadih/aclassdeal/Service/ReportedAidService.java
@@ -13,7 +13,7 @@ public interface ReportedAidService {
 
 	ReportedAdsModel saveAdds(@Valid ReportedAdsModel rAds);
 
-	ReportedAdsModel blockAdds(Long i);
+	ReportedAdsModel blockAdds(int i);
 
 	List<ReportedAdsModel> getallAds(int limit,int offset);
 }
